# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления рекламодателями и их рекламными кампаниями.",
        "title": "Eshke Reklama API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/ad_campaigns/{ad_campaign_id}/ad_groups/{ad_group_id}/ads": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ads"
                ],
                "summary": "Список объявлений группы",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID рекламной кампании",
                        "name": "ad_campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID группы объявлений",
                        "name": "ad_group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListAdsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ads"
                ],
                "summary": "Создание объявления",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID рекламной кампании",
                        "name": "ad_campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID группы объявлений",
                        "name": "ad_group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Заголовок объявления",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Короткое описание объявления",
                        "name": "short_desc",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Целевой URL",
                        "name": "target_url",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение объявления",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            }
        },
        "/api/ad_campaigns/{ad_campaign_id}/ad_groups/{ad_group_id}/ads/{ad_id}": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Пользователь может только включать и выключать объявление; изменение контента повторно отправляет его на модерацию",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ads"
                ],
                "summary": "Обновление объявления",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID рекламной кампании",
                        "name": "ad_campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID группы объявлений",
                        "name": "ad_group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID объявления",
                        "name": "ad_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Новый заголовок объявления",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "turned_off",
                            "working"
                        ],
                        "type": "string",
                        "description": "Новый статус объявления",
                        "name": "status",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Новое короткое описание объявления",
                        "name": "short_desc",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Новый целевой URL",
                        "name": "target_url",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Новое изображение объявления",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ads"
                ],
                "summary": "Удаление объявления",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID рекламной кампании",
                        "name": "ad_campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID группы объявлений",
                        "name": "ad_group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID объявления",
                        "name": "ad_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpx.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "short_desc": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateAdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ListAdsResponse": {
            "type": "object",
            "properties": {
                "ads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AdResponse"
                    }
                },
                "group_id": {
                    "type": "integer"
                }
            }
        },
...
        "httpx.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "httpx.Success": {
            "type": "object",
            "properties": {
                "data": {}
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}